Move bmcutils to pkg/bmcutils
