+++ internal/controller/biossettings_controller_test.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/ironcore-dev/controller-utils/metautils"
 	metalv1alpha1 "github.com/ironcore-dev/metal-operator/api/v1alpha1"
-	"github.com/ironcore-dev/metal-operator/bmcutils"
+	"github.com/ironcore-dev/metal-operator/pkg/bmcutils"
 
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	. "sigs.k8s.io/controller-runtime/pkg/envtest/komega"
