+++ internal/controller/bmc_controller.go
@@ -23,8 +23,8 @@ import (
 	"github.com/ironcore-dev/controller-utils/metautils"
 	metalv1alpha1 "github.com/ironcore-dev/metal-operator/api/v1alpha1"
 	"github.com/ironcore-dev/metal-operator/bmc"
-	"github.com/ironcore-dev/metal-operator/bmcutils"
 	"github.com/ironcore-dev/metal-operator/internal/serverevents"
+	"github.com/ironcore-dev/metal-operator/pkg/bmcutils"
 
 	"github.com/stmcginnis/gofish/schemas"
 	corev1 "k8s.io/api/core/v1"
