+++ internal/controller/bmcsettings_controller_test.go
@@ -18,7 +18,7 @@ import (
 
 	"github.com/ironcore-dev/controller-utils/metautils"
 	metalv1alpha1 "github.com/ironcore-dev/metal-operator/api/v1alpha1"
-	"github.com/ironcore-dev/metal-operator/bmcutils"
+	"github.com/ironcore-dev/metal-operator/pkg/bmcutils"
 )
 
 var _ = Describe("BMCSettings Controller", func() {
